割り込みで読めてそう？

+++ Core/Inc/MMC5983MA.hpp
@@ -37,9 +37,9 @@ private:
 		int16_t z;
 	};
 	struct Gauss{
-		int16_t x;
-		int16_t y;
-		int16_t z;
+		int32_t x;
+		int32_t y;
+		int32_t z;
 	};
 
 	Offset offset_;
